requirements.txt
mysql-connector-python==8.1.0
flask==2.3.3
python_dotenv==1.0.0
flask_cors==4.0.0
